# Optimizations of the Cython Lomb-Scargle periodogram implementations