--- a/astropy/timeseries/periodograms/lombscargle/implementations/cython_impl.pyx
+++ b/astropy/timeseries/periodograms/lombscargle/implementations/cython_impl.pyx
@@ -10,7 +10,7 @@ np.import_array()
 cdef extern from "math.h":
     double sin(double)
     double cos(double)
-    double atan2(double, double)
+    double sqrt(double)
 
 DTYPE = np.float64
 ctypedef np.float64_t DTYPE_t
@@ -20,7 +20,7 @@ ctypedef np.intp_t ITYPE_t
 
 
 def lombscargle_cython(t, y, dy, frequency, normalization='standard',
-                       fit_mean=True, center_data=True):
+                       fit_mean=True, center_data=True, assume_regular_frequency=False):
     """Lomb-Scargle Periodogram
 
     This is a pure-python implementation of the original Lomb-Scargle formalism
@@ -43,6 +43,8 @@ def lombscargle_cython(t, y, dy, frequency, normalization='standard',
     center_data : bool, optional
         if True, pre-center the data by subtracting the weighted mean
         of the input data. This is especially important if ``fit_mean = False``
+    assume_regular_frequency : bool, optional
+        if True, allocate the memory buffers to accelerate the computation
 
     Returns
     -------
@@ -74,15 +76,13 @@ def lombscargle_cython(t, y, dy, frequency, normalization='standard',
 
     # pre-center the data: not technically required if fit_mean=True,
     # but it simplifies the math.
+    w = dy ** -2
+    w /= np.sum(w)
     if fit_mean or center_data:
         # compute MLE for mean in the presence of noise.
-        w = dy ** -2
         y = y - np.dot(w, y) / np.sum(w)
 
-    if fit_mean:
-        _generalized_lomb_scargle(t, y, dy, 2 * np.pi * frequency, PLS)
-    else:
-        _standard_lomb_scargle(t, y, dy, 2 * np.pi * frequency, PLS)
+    _lomb_scargle(t, y, w, 2 * np.pi * frequency, PLS, fit_mean = fit_mean, assume_regular_frequency = assume_regular_frequency)
 
     if normalization == 'standard':
         pass
@@ -98,155 +98,94 @@ def lombscargle_cython(t, y, dy, frequency, normalization='standard',
                          "not recognized".format(normalization))
     return PLS.ravel()
 
-
+from libc.stdlib cimport malloc, free
 @cython.cdivision(True)
 @cython.boundscheck(False)
 @cython.wraparound(False)
-cdef _standard_lomb_scargle(const DTYPE_t[::1] t, const DTYPE_t[::1] y, const DTYPE_t[::1] dy,
-                            const DTYPE_t[::1] omega, DTYPE_t[::1] PLS):
+
+cdef _lomb_scargle(const DTYPE_t[::1] t, const DTYPE_t[::1] y, const DTYPE_t[::1] w,
+                            const DTYPE_t[::1] omega, DTYPE_t[::1] PLS, bint fit_mean=True, bint assume_regular_frequency=False):
+
     cdef ITYPE_t N_freq = omega.shape[0]
     cdef ITYPE_t N_obs = t.shape[0]
 
-    cdef DTYPE_t w, omega_t, sin_omega_t, cos_omega_t
-    cdef DTYPE_t S2, C2, tau, Y, wsum, YY, YCtau, YStau, CCtau, SStau
+    cdef DTYPE_t S, C, Sh, Ch, Sw, Cw, S2, C2, S2w, C2w, YC, YS, YY, tan_2omega_tau
+
+    # allocate buffers used to speed-up the computation for regular grids
+    cdef DTYPE_t* sin_buffer = <DTYPE_t*>malloc(N_obs * sizeof(DTYPE_t))
+    cdef DTYPE_t* cos_buffer = <DTYPE_t*>malloc(N_obs * sizeof(DTYPE_t))
+    cdef DTYPE_t* dsin_buffer = NULL
+    cdef DTYPE_t* dcos_buffer = NULL
+    if assume_regular_frequency:
+        dsin_buffer = <DTYPE_t*>malloc(N_obs * sizeof(DTYPE_t))
+        dcos_buffer = <DTYPE_t*>malloc(N_obs * sizeof(DTYPE_t))
+
+    if assume_regular_frequency:
+        #init the buffers in order to enable recursion
+        for j in range(N_obs):
+            tmp = t[j] * (omega[N_freq-1] - omega[0]) / (N_freq - 1)
+            dsin_buffer[j] = sin(tmp)
+            dcos_buffer[j] = cos(tmp)
+
+    YY = 0
+
+    for j in range(N_obs):
+        YY += y[j] * y[j] * w[j]
 
     for i in range(N_freq):
-        # first pass: determine tau
-        S2 = 0
-        C2 = 0
+        S = C = Sh = Ch = Sw = Cw = S2 = C2 = S2w = C2w = YS = YC = 0
+
+
+        if not assume_regular_frequency or i % 64 == 0 :
+             for j in range(N_obs):
+                sin_buffer[j] = sin(omega[i] * t[j])
+                cos_buffer[j] = cos(omega[i] * t[j])
+        elif assume_regular_frequency :
+            # exp(x + dx) = exp(x) * exp(dx)
+            for j in range(N_obs):
+                tmp = cos_buffer[j] * dcos_buffer[j] - sin_buffer[j] * dsin_buffer[j]
+                sin_buffer[j] = cos_buffer[j] * dsin_buffer[j] + sin_buffer[j] * dcos_buffer[j]
+                cos_buffer[j] = tmp
+
         for j in range(N_obs):
-            w = 1. / dy[j]
-            w *= w
 
-            omega_t = omega[i] * t[j]
-            sin_omega_t = sin(omega_t)
-            cos_omega_t = cos(omega_t)
+            Sh += y[j] * w[j] * sin_buffer[j]
+            Ch += y[j] * w[j] * cos_buffer[j]
+            S2 += w[j] * 2 * sin_buffer[j] * cos_buffer[j]
+            C2 += w[j] * (1 - (2 * sin_buffer[j] * sin_buffer[j]))
 
-            S2 += 2 * w * sin_omega_t * cos_omega_t
-            C2 += w * (1 - 2 * sin_omega_t * sin_omega_t)
+        if fit_mean:
+            for j in range(N_obs):
+                S += w[j] * sin_buffer[j]
+                C += w[j] * cos_buffer[j]
 
-        tau = 0.5 * atan2(S2, C2) / omega[i]
+        if fit_mean:
+            tan_2omega_tau = (S2 - (2 * S * C)) / (C2 - ((C * C) - (S * S)))
+        else:
+            tan_2omega_tau = S2 / C2
 
-        wsum = 0
-        Y = 0
-        YY = 0
-        YCtau = 0
-        YStau = 0
-        CCtau = 0
-        SStau = 0
+        S2w = tan_2omega_tau / sqrt(1 + tan_2omega_tau * tan_2omega_tau)
+        C2w = 1 / sqrt(1 + tan_2omega_tau * tan_2omega_tau)
 
-        # second pass: compute the power
-        for j in range(N_obs):
-            w = 1. / dy[j]
-            w *= w
-            wsum += w
+        Cw = sqrt(0.5 * (1 + C2w))
+        Sw = sqrt(0.5 * (1 - C2w))
+        if S2w < 0:
+            Sw = - Sw
 
-            omega_t = omega[i] * (t[j] - tau)
-            sin_omega_t = sin(omega_t)
-            cos_omega_t = cos(omega_t)
+        YC = Ch * Cw + Sh * Sw
+        YS = Sh * Cw - Ch * Sw
 
-            Y += w * y[j]
-            YY += w * y[j] * y[j]
-            YCtau += w * y[j] * cos_omega_t
-            YStau += w * y[j] * sin_omega_t
-            CCtau += w * cos_omega_t * cos_omega_t
-            SStau += w * sin_omega_t * sin_omega_t
+        CC = 0.5 * (1 + C2 * C2w + S2 * S2w)
+        SS = 0.5 * (1 - C2 * C2w - S2 * S2w)
 
-        Y /= wsum
-        YY /= wsum
-        YCtau /= wsum
-        YStau /= wsum
-        CCtau /= wsum
-        SStau /= wsum
+        if fit_mean:
+            CC -= (C * Cw + S * Sw) * (C * Cw + S * Sw)
+            SS -= (S * Cw - C * Sw) * (S * Cw - C * Sw)
 
-        PLS[i] = (YCtau * YCtau / CCtau + YStau * YStau / SStau) / YY
+        PLS[i] = ((YC * YC / CC) + (YS * YS / SS)) / YY
 
-
-@cython.cdivision(True)
-@cython.boundscheck(False)
-@cython.wraparound(False)
-cdef _generalized_lomb_scargle(const DTYPE_t[::1] t, const DTYPE_t[::1] y, const DTYPE_t[::1] dy,
-                               const DTYPE_t[::1] omega, DTYPE_t[::1] PLS):
-    cdef ITYPE_t N_freq = omega.shape[0]
-    cdef ITYPE_t N_obs = t.shape[0]
-
-    cdef DTYPE_t w, omega_t, sin_omega_t, cos_omega_t
-    cdef DTYPE_t S, C, S2, C2, tau, Y, wsum, YY
-    cdef DTYPE_t Stau, Ctau, YCtau, YStau, CCtau, SStau
-
-    for i in range(N_freq):
-        # first pass: determine tau
-        wsum = 0
-        S = 0
-        C = 0
-        S2 = 0
-        C2 = 0
-        for j in range(N_obs):
-            w = 1. / dy[j]
-            w *= w
-            wsum += w
-
-            omega_t = omega[i] * t[j]
-            sin_omega_t = sin(omega_t)
-            cos_omega_t = cos(omega_t)
-
-            S += w * sin_omega_t
-            C += w * cos_omega_t
-
-            S2 += 2 * w * sin_omega_t * cos_omega_t
-            C2 += w - 2 * w * sin_omega_t * sin_omega_t
-
-        S2 /= wsum
-        C2 /= wsum
-        S /= wsum
-        C /= wsum
-
-        S2 -= (2 * S * C)
-        C2 -= (C * C - S * S)
-
-        tau = 0.5 * atan2(S2, C2) / omega[i]
-
-        Y = 0
-        YY = 0
-        Stau = 0
-        Ctau = 0
-        YCtau = 0
-        YStau = 0
-        CCtau = 0
-        SStau = 0
-
-        # second pass: compute the power
-        for j in range(N_obs):
-            w = 1. / dy[j]
-            w *= w
-
-            omega_t = omega[i] * (t[j] - tau)
-            sin_omega_t = sin(omega_t)
-            cos_omega_t = cos(omega_t)
-
-            Y += w * y[j]
-            YY += w * y[j] * y[j]
-            Ctau += w * cos_omega_t
-            Stau += w * sin_omega_t
-            YCtau += w * y[j] * cos_omega_t
-            YStau += w * y[j] * sin_omega_t
-            CCtau += w * cos_omega_t * cos_omega_t
-            SStau += w * sin_omega_t * sin_omega_t
-
-        Y /= wsum
-        YY /= wsum
-        Ctau /= wsum
-        Stau /= wsum
-        YCtau /= wsum
-        YStau /= wsum
-        CCtau /= wsum
-        SStau /= wsum
-
-        YCtau -= Y * Ctau
-        YStau -= Y * Stau
-        CCtau -= Ctau * Ctau
-        SStau -= Stau * Stau
-
-        YY -= Y * Y
-
-        PLS[i] = (YCtau * YCtau / CCtau + YStau * YStau / SStau) / YY
+    free(sin_buffer)
+    free(cos_buffer)
+    if assume_regular_frequency:
+        free(dsin_buffer)
+        free(dcos_buffer)