<!-- These comments are hidden when you submit the pull request,
so you do not need to remove them! -->

<!-- Please be sure to check out our contributing guidelines,
https://github.com/astropy/astropy/blob/main/CONTRIBUTING.md .
Please be sure to check out our code of conduct,
https://github.com/astropy/astropy/blob/main/CODE_OF_CONDUCT.md . -->

<!-- If you are new or need to be re-acquainted with Astropy
contributing workflow, please see
https://docs.astropy.org/en/latest/development/quickstart.html .
There is even a practical example at
https://docs.astropy.org/en/latest/development/git_edit_workflow_examples.html . -->

<!-- Please just have a quick search on GitHub to see if a similar
pull request has already been posted.
We have old closed pull requests that might provide useful code or ideas
that directly tie in with your pull request. -->

<!-- We have several automatic features that run when a pull request is open.
They can appear daunting but do not worry because maintainers will help
you navigate them, if necessary. -->

### Description
<!-- Provide a general description of what your pull request does.
Complete the following sentence and add relevant details as you see fit. -->

The modifications can be divided into 3 different minor modifications;

1. Both of the Cython methods have been merged and rewritten to avoid redundant recomputation of trigonometric functions.
2. The 'assume_regular_frequency' parameter is passed to the Cython method. If true activates further optimizations limited to regular frequency grids only
3. In the Chi2 implementation and core replace solving the systems of normal equations with dedicated numpy.linalg.lstsq least-squared solver. Slightly improves numerical stability, however in most cases it's only a cosmetic change.

Generally the Scipy implementation after modifications seems to be ~ 50% faster for nonregular grids and ~ 4x  for regular ones. I've not been able to measure any significant discrepancies between results given by the old one and after that (with < 10^-12 discrepancies for MACHO dataset, just like between old Cython and Slow methods), so it shouldn't affect anything else.


<!-- In addition please ensure that the pull request title is descriptive
and allows maintainers to infer the applicable subpackage(s). -->

<!-- READ THIS FOR MANUAL BACKPORT FROM A MAINTAINER:
Apply "skip-basebranch-check" label **before** you open the PR! -->

<!-- Optional opt-out -->