@kkrutkowski 2024-10-23
It looks like the fastchi2 method fails everywhere now. Either lstsq gives wrong results for near-zero values (the biggest absolute deviation seems not to exceed 10^-12) or the checks depend inherently on the instability of systems of normal equations so that errors are reproducible between methods (as fastchi2 is not capable of applying decomposition directly).

Shall I remove that change and limit the changes only to Cython, which seems to be running fine? Alternatively, the atol value for allclose's atol could be increased but this change probably isn't important enough to justify changes like that.