@kkrutkowski 2024-10-24
Sorry for the reply here, but I wasn't able to do this in the right section.

Could you suggest in a little bit more detail what should've I added to the documentation (about the assumptions)? There is already mention of passing this boolean set to true enabling recursive frequency modification.

I mean the modified Cython implementation doesn't require the regular grid, however, if passed 'assume_regular_grid = True) it applies exp^(x + dx) = exp(x) * exp(dx) identity to reduce required calls of trigonometric functions.

If 'assume_regular_grid = False or is not passed at all the method still functions with no issues, but then the speed-up is much more modest (as mentioned in the docstring describing 'assume_regular_grid' in the Cython implementation).