@neutrinoceros 2024-10-24
I guess maybe I was just confused by the following comment: `# enable optimizations specific to regular grids only`
I read it and thought that `method='cython'` only made sense for regular grids, but I think you just clarified that it is not quite so. In that case, it seems that the docstring is indeed complete already, but I'd recommend clarifying that comment.