@kkrutkowski 2024-10-24
Looks like units.time_quantity_times_unit got optimized on main and as such benchmark failed. 

| Change   | Before [bf852938]    | After [54b7f5ed]    |   Ratio | Benchmark (Parameter)          |
|----------|----------------------|---------------------|---------|--------------------------------|
| +        | 46.1±0.2μs           | 99.1±30μs           |    2.15 | units.time_quantity_times_unit |

SOME BENCHMARKS HAVE CHANGED SIGNIFICANTLY.
PERFORMANCE DECREASED.
Error: Process completed with exit code 1

Rerunning the benchmark after synchronization, I hope it'll pass once again (as it did before the comment modification).