@neutrinoceros 2024-10-24
Our benchmarks are unfortunately run on a shared machine, which means results are sometimes a bit flaky.
I added the labels now, but I want to note that we're one day away from astropy 7.0's feature freeze so we will likely not have time to merge this in this cycle, as older PRs are being prioritized and we still have a couple dozens of them. Thanks for your understanding !