@kkrutkowski 2024-10-24
Thanks for the review and reply.

For now, everything looks like the PR is ready to merge, whenever it is the case for 7.0 or 7.1 won't change much I guess.

Also, https://gitlab.gnome.org/GNOME/libxml2 is temporarily down (Error 500), so the 'Documentation link check' can't be completed, but just like the failed benchmark previously it seems unrelated to the changes from this PR. I'll relaunch tests once GitLab is back up (but last time it took them a few days, so I guess it may not be quick).