@pllim 2024-10-24
Re: linkcheck -- If that link is not related to your work, you can ignore. It is flaky most of the time but we keep it because it does catch real broken links sometimes and the way we set up the optional jobs, you get it "for free" with the `Extra CI` label.