@kkrutkowski 2024-10-24
> Re: linkcheck -- If that link is not related to your work, you can ignore. It is flaky most of the time but we keep it because it does catch real broken links sometimes and the way we set up the optional jobs, you get it "for free" with the `Extra CI` label.

Thanks, I think it's ready to merge (or for further modifications if there are any issues), so it'll probably stay like that for some time.

Everything else passed without issues (including all tox tests), so there probably aren't any significant issues with it.