@kkrutkowski 2024-10-25
> I'm only noticing now, but this PR was opened from your `main` branch, which is a problem. Ideally we should fix this now rather than later, because it'll only get worse from here. I'll post-pone my in-depth review to after this problem is addressed. Please have a look at the following article which explains why this should be avoided and what to do next https://hynek.me/articles/pull-requests-branch/ If you need a hand, don't hesitate to reach out.

Thanks, reopened as #17251 now, from a different branch.